# docs/spikes/S7-api-v1/recorded/swagger_v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "MintPlayer API",
    "description": "API that allows access to the data of MintPlayer. ",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/MediumType": {
      "get": {
        "tags": [
          "MediumType"
        ],
        "summary": "Get the medium types from the database.",
        "operationId": "api-mediumtype-list",
        "parameters": [
          {
            "name": "include_relations",
            "in": "header",
            "description": "Specifies whether the related entities should be included in the response.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediumType"
                  }
                }
              },
              "text/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediumType"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediumType"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediumType"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "MediumType"
        ],
        "summary": "Creates a new medium type in the database.",
        "operationId": "api-mediumtype-create",
        "requestBody": {
          "description": "Medium type to be inserted in the database.",
          "content": {
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/MediumType"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/MediumType"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/MediumType"
              }
            },
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/MediumType"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediumType"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MediumType"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MediumType"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/MediumType"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/MediumType"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediumType"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediumType"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/MediumType/{id}": {
      "get": {
        "tags": [
          "MediumType"
        ],
        "summary": "Get a specific medium type from the database.",
        "operationId": "api-mediumtype-get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the medium type to retrieve.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "include_relations",
            "in": "header",
            "description": "Specifies whether the related entities should be included in the response.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/MediumType"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/MediumType"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediumType"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediumType"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "MediumType"
        ],
        "summary": "Updates a medium type in the database.",
        "operationId": "api-mediumtype-update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "New medium type information.",
          "content": {
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/MediumType"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/MediumType"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/MediumType"
              }
            },
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/MediumType"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediumType"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MediumType"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MediumType"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/MediumType"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/MediumType"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediumType"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediumType"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "MediumType"
        ],
        "summary": "Deletes a medium type from the database.",
        "operationId": "api-mediumtype-delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the medium type to delete.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MediumType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "visible": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}